Extract screens to separate directory. Update EncryptionErrorScreen - add warning colors and larger card
